Remove args from Features
`args` and `value` are a duplication for most of the features:
`args = [value]`
get rid of `args` and introduce `value` for `Feature` (the main class
instead of the subclasses). This makes the code simpler.

Refactoring the `freeze_serialize` function to not use an array would
simplify the code a bit more, but that needs a bit more of work.